Separate the types in DownloadManager and MuxOfflineAccessManager into files
