+++ Sources/MuxPlayerSwift/Offline/DownloadManager.swift
@@ -302,361 +302,3 @@ actor DownloadManager {
         downloadTasksByPlaybackID[playbackID] = nil
     }
 }
-
-// MARK: - Internal DTOs
-
-// internal DTO for our index of downloaded assets
-struct StoredAsset: Codable {
-    let isComplete: Bool
-    let completedWithError: Bool
-    
-    let playbackID: String
-    let localPath: String?
-    let readableTitle: String
-    let posterDataBase64: String?
-    let subtitleLanguages: [String]?
-    let secondaryAudioLanguages: [String]?
-    
-    let ckcFilePath: String?
-    /// For DRM keys: Either storage or play expiration depending on state
-    let keyExpiration: Date?
-    /// For secure playback: playback token expiration
-    let redownloadExpiration: Date?
-}
-
-extension StoredAsset {
-    static func forNewDownload(playbackID: String, options: DownloadOptions) -> StoredAsset {
-       return StoredAsset(
-        isComplete: false,
-        completedWithError: false,
-        playbackID: playbackID,
-        localPath: nil,
-        readableTitle: options.readableTitle,
-        posterDataBase64: options.posterData?.base64EncodedString(),
-        subtitleLanguages: options.subtitleLanguages,
-        secondaryAudioLanguages: options.secondaryAudioLanguages,
-        ckcFilePath: nil,
-        keyExpiration: nil,
-        redownloadExpiration: nil
-       )
-    }
-    
-    func debugString() -> String {
-        let encoder = JSONEncoder()
-        encoder.outputFormatting = [.prettyPrinted, .sortedKeys]
-        encoder.dateEncodingStrategy = .iso8601
-        
-        let data: Data
-        do {
-            data = try encoder.encode(self)
-        } catch {
-            return "StoredAsset(playbackID: \(playbackID), encoding failed)"
-        }
-        
-        guard let jsonString = String(data: data, encoding: .utf8) else {
-            return "StoredAsset(playbackID: \(playbackID), encoding failed)"
-        }
-        
-        return jsonString
-    }
-}
-
-extension DownloadOptions {
-    init(from storedAsset: StoredAsset) {
-        self.readableTitle = storedAsset.readableTitle
-        
-        // Decode base64 string to Data
-        if let posterData = storedAsset.posterDataBase64, !posterData.isEmpty {
-            self.posterData = Data(base64Encoded: posterData)
-        } else {
-            self.posterData = nil
-        }
-        
-        self.subtitleLanguages = storedAsset.subtitleLanguages
-        self.secondaryAudioLanguages = storedAsset.secondaryAudioLanguages
-    }
-}
-
-
-// MARK: - DownloadIndex
-
-// Stores a persistent index of downloaded media, along with sidecar data, DRM keys, etc
-actor DownloadIndex {
-    
-    #if DEBUG
-    private let logger = Logger(OSLog(subsystem: "com.mux.player", category: "Mux-Offline"))
-    #else
-    private let logger = Logger(.disabled)
-    #endif
-    
-    static let plistVersion = 1
-    
-    // Versioned snapshot persisted as a binary property list under Application Support
-    private struct IndexSnapshot: Codable {
-        let version: Int
-        let assets: [String: StoredAsset]
-    }
-
-    private var assets: [String: StoredAsset] = [:]
-
-    init() {
-        do {
-            if let snapshot = try loadSnapshot() {
-                assets = snapshot.assets
-            }
-        } catch {
-            logger.error("[Mux-Offline] DownloadIndex init load error: \(error.localizedDescription)")
-        }
-    }
-
-    // MARK: - Public API
-
-    func upsert(_ asset: StoredAsset) {
-        assets[asset.playbackID] = asset
-        persist()
-    }
-
-    func delete(playbackID: String) {
-        assets.removeValue(forKey: playbackID)
-        persist()
-    }
-
-    func get(playbackID: String) -> StoredAsset? {
-        assets[playbackID]
-    }
-
-    func all() -> [StoredAsset] {
-        Array(assets.values)
-    }
-
-    // MARK: - Partial Updates
-
-    func updateIsComplete(playbackID: String, isComplete: Bool, completeWithError: Bool) -> StoredAsset? {
-        // not an error case. Deletion can occur re-entrantly before the delegate callback that calls this
-        guard let existing = assets[playbackID] else {
-            logger.warning("[Mux-Offline] DownloadIndex.updateIsComplete: No existing asset for playbackID \(playbackID)")
-            return nil
-        }
-        let updated = StoredAsset(
-            isComplete: isComplete,
-            completedWithError: completeWithError,
-            playbackID: existing.playbackID,
-            localPath: existing.localPath,
-            readableTitle: existing.readableTitle,
-            posterDataBase64: existing.posterDataBase64,
-            subtitleLanguages: existing.subtitleLanguages,
-            secondaryAudioLanguages: existing.secondaryAudioLanguages,
-            ckcFilePath: existing.ckcFilePath,
-            keyExpiration: existing.keyExpiration,
-            redownloadExpiration: existing.redownloadExpiration
-        )
-        assets[playbackID] = updated
-        persist()
-        
-        return updated
-    }
-
-    func updateCKCFileURL(playbackID: String, ckcFilePath: String) -> StoredAsset? {
-        // not an error case. Deletion can occur re-entrantly before the delegate callback that calls this
-        guard let existing = assets[playbackID] else {
-            logger.warning("[Mux-Offline] DownloadIndex.updateCKCFileURL: No existing asset for playbackID \(playbackID)")
-            return nil
-        }
-        let updated = StoredAsset(
-            isComplete: existing.isComplete,
-            completedWithError: existing.completedWithError,
-            playbackID: existing.playbackID,
-            localPath: existing.localPath,
-            readableTitle: existing.readableTitle,
-            posterDataBase64: existing.posterDataBase64,
-            subtitleLanguages: existing.subtitleLanguages,
-            secondaryAudioLanguages: existing.secondaryAudioLanguages,
-            ckcFilePath: ckcFilePath,
-            keyExpiration: existing.keyExpiration,
-            redownloadExpiration: existing.redownloadExpiration
-        )
-        assets[playbackID] = updated
-        persist()
-        
-        return updated
-    }
-
-    func updateLocalPathURL(playbackID: String, localPath: String) -> StoredAsset? {
-        // not an error case. Deletion can occur re-entrantly before the delegate callback that calls this
-        guard let existing = assets[playbackID] else {
-            logger.warning("[Mux-Offline] DownloadIndex.updateLocalPathURL: No existing asset for playbackID \(playbackID)")
-            return nil
-        }
-        let updated = StoredAsset(
-            isComplete: existing.isComplete,
-            completedWithError: existing.completedWithError,
-            playbackID: existing.playbackID,
-            localPath: localPath,
-            readableTitle: existing.readableTitle,
-            posterDataBase64: existing.posterDataBase64,
-            subtitleLanguages: existing.subtitleLanguages,
-            secondaryAudioLanguages: existing.secondaryAudioLanguages,
-            ckcFilePath: existing.ckcFilePath,
-            keyExpiration: existing.keyExpiration,
-            redownloadExpiration: existing.redownloadExpiration
-        )
-        assets[playbackID] = updated
-        persist()
-        return updated
-    }
-
-    // MARK: - Persistence
-
-    private func persist() {
-        do {
-            try saveSnapshot(IndexSnapshot(version: Self.plistVersion, assets: assets))
-        } catch {
-            logger.error("[Mux-Offline] Failed to save index snapshot: \(error.localizedDescription)")
-        }
-    }
-
-    private func indexDirectoryURL() throws -> URL {
-        let fm = FileManager.default
-        let base = try fm.url(for: .applicationSupportDirectory, in: .userDomainMask, appropriateFor: nil, create: true)
-        let dir = base.appendingPathComponent("com.mux.offline", isDirectory: true)
-        try fm.createDirectory(at: dir, withIntermediateDirectories: true)
-        // Exclude from iCloud backups
-        var values = URLResourceValues()
-        values.isExcludedFromBackup = true
-        var mutableDir = dir
-        try? mutableDir.setResourceValues(values)
-        return dir
-    }
-
-    private func indexFileURL() throws -> URL {
-        try indexDirectoryURL().appendingPathComponent("index.plist", isDirectory: false)
-    }
-
-    private func saveSnapshot(_ snapshot: IndexSnapshot) throws {
-        let encoder = PropertyListEncoder()
-        encoder.outputFormat = .binary
-        let data = try encoder.encode(snapshot)
-
-        let url = try indexFileURL()
-        let tmp = url.deletingLastPathComponent().appendingPathComponent(UUID().uuidString)
-        try data.write(to: tmp, options: .atomic)
-        do {
-            _ = try FileManager.default.replaceItemAt(url, withItemAt: tmp)
-        } catch {
-            try data.write(to: url)
-        }
-    }
-
-    private func loadSnapshot() throws -> IndexSnapshot? {
-        let url = try indexFileURL()
-        guard FileManager.default.fileExists(atPath: url.path) else { return nil }
-        let data = try Data(contentsOf: url)
-        return try PropertyListDecoder().decode(IndexSnapshot.self, from: data)
-    }
-}
-
-// MARK: - Extensions
-
-extension DownloadEvent {
-    func debugString() -> String {
-        switch self {
-        case .started:
-            return "started"
-        case .waitingForConnectivity:
-            return "waitingForConnectivity"
-        case .progress(let percent):
-            return "progress(\(String(format: "%.1f", percent))%)"
-        case .completed(let asset):
-            return "completed(\(asset.avAssetIfPlayable()?.url.lastPathComponent ?? "unknown"))"
-        }
-    }
-}
-
-// MARK: - Combine to AsyncStream Bridge
-
-extension AnyPublisher {
-    func toAsyncThrowingStream() -> AsyncThrowingStream<Output, Error> {
-        AsyncThrowingStream { continuation in
-            let cancellable = self.sink(
-                receiveCompletion: { completion in
-                    switch completion {
-                    case .finished:
-                        continuation.finish()
-                    case .failure(let error):
-                        continuation.finish(throwing: error)
-                    }
-                },
-                receiveValue: { value in
-                    continuation.yield(value)
-                }
-            )
-            
-            continuation.onTermination = { @Sendable _ in
-                cancellable.cancel()
-            }
-        }
-    }
-}
-
-// MARK: - DownloadTaskDelegate
-
-fileprivate class DownloadTaskDelegate: NSObject, AVAssetDownloadDelegate {
-    #if DEBUG
-    private let logger = Logger(OSLog(subsystem: "com.mux.player", category: "Mux-Offline"))
-    #else
-    private let logger = Logger(.disabled)
-    #endif
-
-    let downloadManager: DownloadManager
-    
-    init(offlineAccessManager: DownloadManager) {
-        self.downloadManager = offlineAccessManager
-    }
-
-    // MARK: AVAssetDownloadDelegate
-
-    func urlSession(_ session: URLSession, task: URLSessionTask, didCompleteWithError error: (any Error)?) {
-        logger.error("[Mux-Offline] didCompleteWithError: taskId=\(task.taskIdentifier) error=\(String(describing: error))")
-        // note - AVAssetDownloadDelegate has a seprarate delegate call for task success, didFinishDownloadingTo, which we use instead of reporting it here
-        if let error {
-            Task { [downloadManager] in
-                await downloadManager.handleError(for: task, error: error)
-            }
-        }
-    }
-
-    func urlSession(_ session: URLSession, assetDownloadTask: AVAssetDownloadTask, didFinishDownloadingTo location: URL) {
-        logger.log("[Mux-Offline] didFinishDownloadingTo: taskId=\(assetDownloadTask.taskIdentifier) location=\(location.path)")
-        Task { [downloadManager] in
-            await downloadManager.handleFinishedDownload(task: assetDownloadTask, location: location)
-        }
-    }
-    
-    // Called when a task becomes a download task (general URLSession delegate)
-    func urlSession(_ session: URLSession, taskIsWaitingForConnectivity task: URLSessionTask) {
-        logger.log("[Mux-Offline] Task waiting for connectivity: id=\(task.taskIdentifier)")
-        Task { [downloadManager] in
-            await downloadManager.handleWaitingForConnectivity(for: task)
-        }
-    }
-
-    func urlSession(_ session: URLSession, assetDownloadTask: AVAssetDownloadTask,
-                    didLoad timeRange: CMTimeRange, totalTimeRangesLoaded loadedTimeRanges: [NSValue],
-                    timeRangeExpectedToLoad: CMTimeRange) {
-        let loaded = loadedTimeRanges.map { $0.timeRangeValue }
-        let loadedDescription = loaded.map { "[start: \($0.start.seconds), dur: \($0.duration.seconds)]" }.joined(separator: ", ")
-        logger.trace("[Mux-Offline] didLoad timeRange start=\(timeRange.start.seconds) dur=\(timeRange.duration.seconds) expectedDur=\(timeRangeExpectedToLoad.duration.seconds) loaded=\(loadedDescription)")
-        Task { [downloadManager] in
-            await downloadManager.handleProgress(task: assetDownloadTask, loadedTimeRanges: loaded, expectedTimeRange: timeRangeExpectedToLoad)
-        }
-    }
-
-    func urlSession(_ session: URLSession, assetDownloadTask: AVAssetDownloadTask,
-                    willDownloadTo location: URL) {
-        logger.log("[Mux-Offline] willDownloadTo [Relative]: \(location.relativePath)")
-        Task { [downloadManager] in
-            await downloadManager.handleDownloadLocation(task: assetDownloadTask, relativeLocation: location.relativePath)
-        }
-    }
-}
-
