+++ Sources/MuxPlayerSwift/PublicAPI/Offline/MuxOfflineAccessManager.swift
@@ -121,66 +121,4 @@ public class MuxOfflineAccessManager {
     }
 }
 
-// MARK: - Public Types
-
-/// Options for configuring a download
-public struct DownloadOptions {
-    /// A human-readable title for the download
-    public let readableTitle: String
-    /// Optional poster image data
-    public let posterData: Data?
-    /// Language codes for subtitles (e.g., 'en' or 'en-US')
-    public let subtitleLanguages: [String]?
-    /// Language codes for secondary audio tracks (e.g., 'en' or 'en-US')
-    public let secondaryAudioLanguages: [String]?
-    
-    public init(readableTitle: String) {
-        self.readableTitle = readableTitle
-        
-        self.posterData = nil
-        self.subtitleLanguages = nil
-        self.secondaryAudioLanguages = nil
-    }
-}
-
-/// Events emitted during a download
-public enum DownloadEvent {
-    /// The download has started
-    case started
-    /// The download is waiting for network connectivity
-    case waitingForConnectivity
-    /// Progress update with percentage complete
-    case progress(percent: Double)
-    /// The download has completed successfully
-    case completed(DownloadedAsset)
-}
-
-/// A downloaded video asset
-public struct DownloadedAsset {
-    /// The Mux playback ID
-    public let playbackID: String
-    /// The current status of the asset
-    public let assetStatus: AssetStatus
-    /// The options used when downloading
-    public let downloadOptions: DownloadOptions
-    
-    /// Returns the AVURLAsset if the asset is playable, nil otherwise
-    public func avAssetIfPlayable() -> AVURLAsset? {
-        if case .playable(asset: let asset) = assetStatus {
-            return asset
-        } else {
-            return nil
-        }
-    }
-}
-
-/// The status of a downloaded asset
-public enum AssetStatus {
-    /// The asset is ready to play
-    case playable(asset: AVURLAsset)
-    /// The asset needs to be re-downloaded when online
-    case redownloadWhenOnline
-    /// The asset's playback period has expired. This value is only applicable to DRM-protected content
-    case expired
-}
